rename to federatedgraph
